Varios cambios para mejorar la seguridad

+++ api/admin/admin.controller.js
@@ -1,19 +1,13 @@
 const adminModel = require("./admin.model");
 const mongoose = require("mongoose");
-const bcrypt = require("bcrypt");
-const axios = require("axios");
 
 module.exports = {
   getAllAdmins,
   getById,
   editPatch,
-  deleteUser,
+  editSelf,
   deleteSelfAdmin,
 };
-
-function sendAdminId(id) {
-  return axios.get("localhost:3010/newadmin/" + id).then((r) => console.log(r));
-}
 
 function getAllAdmins(req, res) {
   if (req.currentUser.role === "admin") {
@@ -63,14 +57,27 @@ function editPatch(req, res) {
   }
 }
 
-function deleteUser(req, res) {
+function editSelf(req, res) {
   if (req.currentUser.role === "admin") {
-    let userId = mongoose.Types.ObjectId.isValid(req.params.id);
-    if (userId) {
+    let adminId = mongoose.Types.ObjectId.isValid(req.params.id);
+    if (adminId) {
       adminModel
-        .findByIdAndDelete(req.params.id)
-        .then((r) => res.json(r))
-        .catch((err) => res.status(500).json("Error en la base de datos"));
+        .findById(req.params.id)
+        .then((r) => {
+          if (r.email === req.currentUser.email) {
+            adminModel
+              .findByIdAndUpdate(req.params.id, req.body)
+              .then((r) => res.json("Eliminado con exito"))
+              .catch((err) =>
+                res.status(500).json("Error en la base de datos")
+              );
+          } else {
+            res
+              .status(400)
+              .send("El email actual no coincide con el email a eliminar");
+          }
+        })
+        .catch((err) => send.status(404).send("Ningun admin con ese ID"));
     } else {
       res.status(404).send("Ningun usuario encontrado");
     }
@@ -80,24 +87,28 @@ function deleteUser(req, res) {
 }
 
 function deleteSelfAdmin(req, res) {
-  let adminId = mongoose.Types.ObjectId.isValid(req.params.id);
+  if (req.currentUser.role === "admin") {
+    let adminId = mongoose.Types.ObjectId.isValid(req.params.id);
 
-  if (adminId) {
-    adminModel
-      .findById(req.params.id)
-      .then((r) => {
-        if (r.email === req.currentUser.email) {
-          adminModel
-            .findByIdAndDelete(req.params.id)
-            .then((r) => res.send("Eliminado con exito"));
-        } else {
-          res
-            .status(403)
-            .send("El email actual no coincide con el email a eliminar");
-        }
-      })
-      .catch((err) => res.status(500).json("Error en la base de datos"));
+    if (adminId) {
+      adminModel
+        .findById(req.params.id)
+        .then((r) => {
+          if (r.email === req.currentUser.email) {
+            adminModel
+              .findByIdAndDelete(req.params.id)
+              .then((r) => res.send("Eliminado con exito"));
+          } else {
+            res
+              .status(403)
+              .send("El email actual no coincide con el email a eliminar");
+          }
+        })
+        .catch((err) => send.status(404).send("Ningun admin con ese ID"));
+    } else {
+      res.status(404).send("Ningun usuario encontrado");
+    }
   } else {
-    res.status(404).send("Ningun usuario encontrado");
+    res.status(403).send("No eres admin");
   }
 }
